requirements.txt
streamlit>=1.35.0
pdfplumber>=0.11.0
pymupdf
requests
python-dotenv
